Feat: Comment 컴포넌트 분리
